Rework shutdown responsibilities of default chain, chain, and cached providers
