+++ source/credentials_provider_cached.c
@@ -24,7 +24,6 @@ AWS_STATIC_STRING_FROM_LITERAL(s_credential_expiration_env_var, "AWS_CREDENTIAL_
 
 struct aws_credentials_provider_cached {
     struct aws_credentials_provider *source;
-    struct aws_credentials_provider_shutdown_options source_shutdown_options;
     struct aws_credentials *cached_credentials;
     struct aws_mutex lock;
     uint64_t refresh_interval_in_ns;
@@ -223,25 +222,6 @@ static void s_cached_credentials_provider_destroy(struct aws_credentials_provide
 
     aws_credentials_provider_release(impl->source);
 
-    /* Clean up memory, mutex, credentials, etc... in the shutdown callback below */
-}
-
-static void s_on_credentials_provider_shutdown(void *user_data) {
-    struct aws_credentials_provider *provider = user_data;
-    if (provider == NULL) {
-        return;
-    }
-
-    struct aws_credentials_provider_cached *impl = provider->impl;
-    if (impl == NULL) {
-        return;
-    }
-
-    /* The wrapped provider has shut down, invoke its shutdown callback if there was one */
-    if (impl->source_shutdown_options.shutdown_callback != NULL) {
-        impl->source_shutdown_options.shutdown_callback(impl->source_shutdown_options.shutdown_user_data);
-    }
-
     /* Invoke our own shutdown callback */
     aws_credentials_provider_invoke_shutdown_callback(provider);
 
@@ -319,13 +299,6 @@ struct aws_credentials_provider *aws_credentials_provider_new_cached(
     } else {
         impl->system_clock_fn = &aws_sys_clock_get_ticks;
     }
-
-    /*
-     * Save the wrapped provider's shutdown callback and then swap it with our own.
-     */
-    impl->source_shutdown_options = impl->source->shutdown_options;
-    impl->source->shutdown_options.shutdown_callback = s_on_credentials_provider_shutdown;
-    impl->source->shutdown_options.shutdown_user_data = provider;
 
     provider->shutdown_options = options->shutdown_options;
